@@ -161,10 +161,10 @@ class LineChart extends StatelessWidget {
           argumentsLabels: argumentsLabels,
           values: values,
           valuesLabels: valuesLabels,
-          horizontalLabelsTextStyle:
-              horizontalLabelsTextStyle ?? Theme.of(context).textTheme.caption,
+          horizontalLabelsTextStyle: horizontalLabelsTextStyle ??
+              Theme.of(context).textTheme.bodySmall,
           verticalLabelsTextStyle:
-              verticalLabelsTextStyle ?? Theme.of(context).textTheme.caption,
+              verticalLabelsTextStyle ?? Theme.of(context).textTheme.bodySmall,
           horizontalLinesPaint: horizontalLinesPaint,
           verticalLinesPaint: verticalLinesPaint,
           additionalMinimalHorizontalLabelsInterval:
